requirements.txt
altgraph==0.16.1
appdirs==1.4.3
aspy.yaml==1.3.0
atomicwrites==1.3.0
attrs==19.1.0
cached-property==1.5.1
Cerberus==1.3.1
certifi==2019.6.16
cfgv==2.0.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
distlib==0.2.9.post0
docopt==0.6.2
entrypoints==0.3
ffmpeg-python==0.2.0
first==2.0.2
flake8==3.7.8
future==0.17.1
identify==1.4.5
idna==2.8
importlib-metadata==0.18
isort==4.3.21
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.3
numpy==1.16.4
orderedmultidict==1.0.1
packaging==19.0
pefile==2019.4.18
pep517==0.5.0
Pillow==6.1.0
pip-shims==0.3.3
pipreqs==0.4.9
plette==0.2.2
pluggy==0.12.0
pre-commit==1.17.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyInstaller==3.5
pyparsing==2.4.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pytest==5.0.1
python-vlc==3.0.6109
pytoml==0.1.21
pywin32-ctypes==0.2.0
PyYAML==5.1.1
QtAwesome==0.5.8
QtPy==1.9.0
requests==2.22.0
requirementslib==1.5.3
six==1.12.0
toml==0.10.0
tomlkit==0.5.5
urllib3==1.25.3
virtualenv==16.7.0
vistir==0.4.3
wcwidth==0.1.7
yarg==0.1.9
zipp==0.5.2
